pyproject.toml
[project]
name = "redesneuronalesavanzadas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tensorflow-datasets>=4.9.9",
]
